Purged CTest & Standardized LibJournal's Unit Test[s]
As much as we love the idea of incorperating testing into the build system,
and as powerful as CTest is, CTest leaves far too much to be desired when
mangling it with other testing tools like BCUnit.

BCUnit is prefered for it's power and simplicity.  And if we ever stop using
BCUnit for any reason, we will end up creating out own such tool in it's likeness.

However, it is understood that BCUnit is really a Unit Testing Toolkit, and that
we may need to explore other options when we get to the Integration Testing stage.

Nevermind the other pitfall of BCUnit, that it's Automated Interface creates
an insane XML page.  But that at least looks reparable by custom building our own
interface via the BCUnit plumbing.  (That BCUnit publicly documents it's plumbing
like that is one reason it is preffered.)